+++ app/components/boot.js
@@ -12,6 +12,8 @@ import 'jquery';
 import 'bootstrap/js/bootstrap';
 import 'bootstrap/css/bootstrap.css!';
 
+import 'font-awesome/css/font-awesome.css!';
+
 import angular from 'angular';
 import app from './app';