+++ dataset/example/gulp/config.js
@@ -1,8 +1,5 @@
 export default {
   deploy: {
-    ghPages: {
-      remoteUrl: "git@github.com:Hypercubed/Project-Chi-Test.git",
-      branch: "gh-pages"
-    }
+    // empty to avoid accidental deploy
   }
 };